@@ -2,57 +2,58 @@
   "name": "client",
   "version": "0.3.23",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "prettier . --check && eslint .",
     "test:e2e": "playwright test"
   },
   "dependencies": {
-    "@auth/drizzle-adapter": "^1.4.1",
-    "@sentry/nextjs": "^8.20.0",
-    "@tanstack/react-query": "^5.51.15",
-    "@upstash/ratelimit": "^2.0.1",
-    "@upstash/redis": "^1.33.0",
-    "@vercel/analytics": "^1.3.1",
-    "@xstate/react": "^4.1.1",
-    "autoprefixer": "10.4.19",
+    "@auth/drizzle-adapter": "^1.11.1",
+    "@sentry/nextjs": "^10.29.0",
+    "@tailwindcss/postcss": "^4.1.17",
+    "@tanstack/react-query": "^5.90.12",
+    "@vercel/analytics": "^1.6.1",
+    "@xstate/react": "^6.0.0",
+    "autoprefixer": "10.4.22",
     "clsx": "^2.1.1",
     "database": "workspace:*",
-    "drizzle-orm": "^0.32.1",
-    "framer-motion": "^11.3.17",
-    "next": "^14.2.5",
-    "next-auth": "^4.24.7",
-    "postcss": "8.4.40",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-hot-toast": "^2.4.1",
-    "react-icons": "^5.2.1",
+    "drizzle-orm": "^0.45.0",
+    "framer-motion": "^12.23.25",
+    "import-in-the-middle": "^2.0.0",
+    "next": "^16.0.7",
+    "next-auth": "^4.24.13",
+    "postcss": "8.5.6",
+    "react": "^19.2.1",
+    "react-dom": "^19.2.1",
+    "react-hot-toast": "^2.6.0",
+    "react-icons": "^5.5.0",
+    "require-in-the-middle": "^8.0.1",
     "server-only": "^0.0.1",
-    "sharp": "^0.33.4",
-    "socket.io-client": "^4.7.5",
-    "tailwind-merge": "^2.4.0",
-    "tailwindcss": "^3.4.7",
-    "typescript": "5.4.4",
-    "use-sound": "^4.0.3",
-    "xstate": "^5.16.0"
+    "sharp": "^0.34.5",
+    "socket.io-client": "^4.8.1",
+    "tailwind-merge": "^3.4.0",
+    "tailwindcss": "^4.1.17",
+    "typescript": "5.9.3",
+    "use-sound": "^5.0.0",
+    "xstate": "^5.24.0"
   },
   "devDependencies": {
-    "@playwright/test": "^1.45.3",
-    "@tailwindcss/typography": "^0.5.13",
-    "@types/node": "^20.14.12",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
+    "@playwright/test": "^1.57.0",
+    "@tailwindcss/typography": "^0.5.19",
+    "@types/node": "^24.10.1",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
     "eslint-config-custom": "workspace:*",
-    "prettier-plugin-tailwindcss": "^0.6.5",
     "tsconfig": "workspace:*"
   },
   "engines": {
-    "node": "20.16.0"
+    "node": "22.14.0"
   },
   "optionalDependencies": {
-    "bufferutil": "^4.0.8",
-    "utf-8-validate": "^6.0.4"
+    "bufferutil": "^4.0.9",
+    "utf-8-validate": "^6.0.5"
   }
 }
