@@ -8,14 +8,15 @@
     "format": "prettier --write \"**/*.{ts,tsx,md}\""
   },
   "devDependencies": {
-    "eslint": "8.57.0",
-    "prettier": "^3.3.3",
+    "eslint": "9.39.1",
+    "prettier": "^3.7.4",
+    "prettier-plugin-tailwindcss": "^0.7.2",
     "tsconfig": "workspace:*",
-    "turbo": "^2.0.9"
+    "turbo": "^2.6.3"
   },
   "engines": {
-    "node": "20.16.0"
+    "node": "22.14.0"
   },
-  "packageManager": "pnpm@9.6.0",
+  "packageManager": "pnpm@10.24.0",
   "name": "artificial-unintelligence"
 }
