@@ -6,20 +6,22 @@ import Lobby from "./lobby";
 import { authOptions } from "@ai/pages/api/auth/[...nextauth]";
 import { isUserInRoom } from "@ai/utils/server-actions";
 
-export default async function Room({ params }: { params: { code: string } }) {
+export default async function Room({ params }: PageProps<"/room/[code]">) {
   const session = await getServerSession(authOptions());
+
+  const { code } = await params;
 
   if (!session) {
     redirect("/");
   }
 
-  const userInRoom = await isUserInRoom({ roomCode: params.code, session });
+  const userInRoom = await isUserInRoom({ roomCode: code, session });
 
   if (!userInRoom) {
     redirect("/");
   }
 
-  const roomInfo = await getRoomInfo(params.code);
+  const roomInfo = await getRoomInfo(code);
 
   return <Lobby roomInfo={roomInfo} session={session} />;
 }
