@@ -30,7 +30,9 @@ const Winner = ({ leaderboard }: WinnerProps) => {
     }
 
     return () => {
-      controls && controls.stop();
+      if (controls) {
+        controls.stop();
+      }
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
