@@ -9,7 +9,7 @@ type HeaderProps = {
 
 const Header = ({ session }: HeaderProps) => {
   return (
-    <header className="container absolute left-0 right-0 top-4 z-50 mx-auto mt-4 flex items-center justify-between gap-4 px-4 md:top-8">
+    <header className="absolute top-4 right-0 left-0 z-50 container mx-auto mt-4 flex items-center justify-between gap-4 px-4 md:top-8">
       <Link href="/" className="flex">
         <Friend className="mr-2 w-8" />
         Aritificial Unintelligence
