@@ -14,7 +14,9 @@ const useClickAway = (
   useEffect(() => {
     const handler = (event: Event) => {
       const { current: el } = ref;
-      el && !el.contains(event.target as Node) && savedCallback.current(event);
+      if (el && !el.contains(event.target as Node)) {
+        savedCallback.current(event);
+      }
     };
     for (const eventName of events) {
       document.addEventListener(eventName, handler);
