@@ -2,7 +2,7 @@ import { drizzle } from "drizzle-orm/postgres-js";
 import postgres from "postgres";
 
 const client = postgres(
-  process.env.POSTGRES_URL ?? process.env.DATABASE_URL ?? ""
+  process.env.POSTGRES_URL ?? process.env.DATABASE_URL ?? "",
 );
 
 export const db = drizzle(client);
